# src/main/resources/swagger2.json
{
	"swagger": "2.0",
	"info": {
		"description": "Spring Boot Swagger Example Api Creator",
		"version": "1.0",
		"title": "Spring Boot Swagger Example API",
		"termsOfService": "Terms of Service",
		"contact": {
			"name": "yourname",
			"url": "https://github.com/yourlink",
			"email": "youremail@yourdomain.com"
		},
		"license": {
			"name": "Apache License Version 2.0",
			"url": "https://www.apache.org/licesen.html"
		}
	},
	"host": "210.117.172.165:8085",
	"basePath": "/",
	"schemes":"http",
	"tags": [{
		"name": "rest-user-controller",
		"description": "Rest User Controller"
	}],
	"paths": {
		"/api/v1/authenticate": {
			"post": {
				"tags": ["rest-user-controller"],
				"summary": "createAuthenticationToken",
				"operationId": "createAuthenticationTokenUsingPOST",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"in": "body",
					"name": "authenticationRequest",
					"description": "authenticationRequest",
					"required": true,
					"schema": {
						"$ref": "#/definitions/AuthenticationRequest"
					}
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/v1/login": {
			"post": {
				"tags": ["rest-user-controller"],
				"summary": "signIn",
				"operationId": "signInUsingPOST",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "id",
					"in": "query",
					"required": false,
					"type": "integer",
					"format": "int32"
				}, {
					"name": "role",
					"in": "query",
					"required": false,
					"type": "integer",
					"format": "int32"
				}, {
					"name": "login",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "password",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "username",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "role_name",
					"in": "query",
					"required": false,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Result"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/v1/member/info": {
			"get": {
				"tags": ["rest-user-controller"],
				"summary": "memberInfo",
				"operationId": "memberInfoUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "id",
					"in": "query",
					"required": false,
					"type": "integer",
					"format": "int32"
				}, {
					"name": "role",
					"in": "query",
					"required": false,
					"type": "integer",
					"format": "int32"
				}, {
					"name": "login",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "password",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "username",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "role_name",
					"in": "query",
					"required": false,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Result"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/v1/signup": {
			"post": {
				"tags": ["rest-user-controller"],
				"summary": "signUp",
				"operationId": "signUpUsingPOST",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "id",
					"in": "query",
					"required": false,
					"type": "integer",
					"format": "int32"
				}, {
					"name": "role",
					"in": "query",
					"required": false,
					"type": "integer",
					"format": "int32"
				}, {
					"name": "login",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "password",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "username",
					"in": "query",
					"required": false,
					"type": "string"
				}, {
					"name": "role_name",
					"in": "query",
					"required": false,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Result"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"AuthenticationRequest": {
			"type": "object",
			"properties": {
				"password": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			}
		},
		"Result": {
			"type": "object",
			"properties": {
				"data": {
					"type": "object"
				},
				"message": {
					"type": "string"
				},
				"statusCode": {
					"type": "integer",
					"format": "int32"
				},
				"totalCount": {
					"type": "integer",
					"format": "int32"
				}
			}
		}
	}
}